+++ packages/reactive_forms_annotations/lib/reactive_forms_annotations.dart
@@ -1,6 +1,7 @@
 library reactive_forms_annotations;
 
 export 'package:flutter/widgets.dart';
+export 'package:reactive_forms/reactive_forms.dart';
 export 'src/reactive_form_annotation.dart';
 export 'src/form_group_annotation.dart';
 export 'src/form_array_annotation.dart';