+++ packages/reactive_forms_generator/example/lib/docs/freezed2/test.dart
@@ -1,4 +1,3 @@
-import 'package:reactive_forms/reactive_forms.dart';
 import 'package:dartz/dartz.dart' as dartz;
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:reactive_forms_annotations/reactive_forms_annotations.dart';