+++ packages/reactive_forms_generator/example/lib/docs/login_extended_nullable/login_extended_nullable.gform.dart
@@ -592,6 +592,10 @@ class LoginExtendedNullableForm implements FormModel<LoginExtendedNullable> {
         height: heightValue);
   }
 
+  LoginExtendedNullableForm copyWithPath(String? path) {
+    return LoginExtendedNullableForm(loginExtendedNullable, form, path);
+  }
+
   void updateValue(LoginExtendedNullable value,
           {bool updateParent = true, bool emitEvent = true}) =>
       form.updateValue(