+++ packages/reactive_forms_generator/example/lib/docs/mailing_list/mailing_list.dart
@@ -1,4 +1,3 @@
-import 'package:reactive_forms/reactive_forms.dart';
 import 'package:example/helpers.dart';
 import 'package:reactive_forms_annotations/reactive_forms_annotations.dart';