+++ packages/reactive_forms_generator/example/lib/docs/user_profile/user_profile.gform.dart
@@ -357,6 +357,10 @@ class UserProfileForm implements FormModel<UserProfile> {
         office: officeValue);
   }
 
+  UserProfileForm copyWithPath(String? path) {
+    return UserProfileForm(userProfile, form, path);
+  }
+
   void updateValue(UserProfile value,
           {bool updateParent = true, bool emitEvent = true}) =>
       form.updateValue(
@@ -601,6 +605,10 @@ class AddressForm implements FormModel<Address> {
       );
     }
     return Address(street: streetValue, city: cityValue, zip: zipValue);
+  }
+
+  AddressForm copyWithPath(String? path) {
+    return AddressForm(address, form, path);
   }
 
   void updateValue(Address value,